SERVER-29237 add maxConnecting in asio connpool
A new connection pool option which controls that rate at which we add
new connections.  It changes behavior so that only N connections can be
in the processing state at any one time (in setup/refresh).  More
connections will be added, if needed, as each new connection completes,
fails or times out.

available in mongos via ShardingTaskExecutorPoolMaxConnecting, defaults
to unlimited.