SERVER-40166 Force BG clock now >= Date_t::lastNow
Currently our fast clock is implemented by a background thread which
wakes up every 10ms if callers continue to check it's time. If that
background thread get's descheduled for a long time for some reason,
this can cause the background threads time to drift substantially.

If we force every read of date_t::now to update a new date_t::lastNow,
and force the background clock now() to return the greater of it's own
notion of time and date_t::lastNow, we can ensure that the background
clock is no staler than the last call to Date_t::now